feat: Refactor BasicAuth middleware and related classes
The commit renames and moves the `BasicAuth` middleware and related classes to the `Support\BasicAuth` namespace. This improves the organization and structure of the codebase. The `BasicAuth` middleware is now located in the `Support\BasicAuth` directory, and the `BasicAuthFactory`, `BasicAuthorizedException`, and `BasicAuthenticator` classes are also moved to the same directory. Additionally, the `BasicAuth` middleware now uses the `BasicAuthenticator` class for authentication.
