+++ src/CmsToolProvider.php
@@ -18,8 +18,8 @@ use Takemo101\Chubby\Config\ConfigRepository;
 use Takemo101\Chubby\Filesystem\LocalFilesystem;
 use Takemo101\Chubby\Hook\Hook;
 use Takemo101\Chubby\Http\GlobalMiddlewareCollection;
-use Takemo101\CmsTool\Http\Middleware\BasicAuth;
-use Takemo101\CmsTool\Http\Middleware\BasicAuthFactory;
+use Takemo101\CmsTool\Support\BasicAuth\BasicAuth;
+use Takemo101\CmsTool\Support\BasicAuth\BasicAuthFactory;
 use Takemo101\CmsTool\Http\Middleware\CacheControl;
 use Takemo101\CmsTool\Http\Session\AdminSessionFactory;
 use Takemo101\CmsTool\Http\Session\DefaultAdminSessionFactory;
