+++ src/Support/BasicAuth/BasicAuth.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Takemo101\CmsTool\Http\Middleware;
+namespace Takemo101\CmsTool\Support\BasicAuth;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Takemo101\CmsTool\Http\Exception\BasicAuthorizedException;
+use Takemo101\CmsTool\Support\BasicAuth\BasicAuthorizedException;
 use Takemo101\CmsTool\Support\BasicAuth\BasicAuthenticator;
 use Takemo101\CmsTool\Support\BasicAuth\BasicAuthHeaderParser;
